requirements.txt
grpcio==1.65.5
grpcio-tools==1.65.5
protobuf==5.28.2
